feat: Enable auth without credential files
With this commit, the `auth` crate will read credentials from
the `GOOGLE_APPLICATION_CREDENTIALS_JSON` env variable if its set,
making it possible to authorize google-cloud in environments without
filesystems, or where it doesn't make sense to tie down
auth to files.
For maximum backwards compatibility, signatures
are left untouched and keeping true to the original
spirit the API surface is not increased beyond the
new environment variable.