Use project ID specified in config  for Pubsub client (#243)
If the Pubsub client tries to access the pubsub topic or subscription of a
project different from the service account's project, the TokenSource's
project_id (the client's service account's project ID) will be set, so the
topic and subscription of the external project will be set. cannot be
accessed.
Therefore, if the project ID is specified in the config, use that.