storage: make size and etag fields optional
These fields are marked as `omitempty` in the Google Cloud Storage's
Go SDK. As a result, when an object with size == 0 is encoded using
the Go SDK's *storage.Object representation it cannot be parsed by
this crate.

Fix this by defaulting the size and etag fields to
zero/the empty string, respectively.