Fix job statistics deserialization for cached queries
The `QUERY_CACHE_HIT` variant was missing from the IndexUnusedCode
enum. This caused deserialization to fail when requesting job details
for a query that's been cached by Google Cloud.

Additionally, this commit also adds a second variant that was missing
from this enum, `STALE_INDEX`.

See the Google Cloud docs for reference:
https://cloud.google.com/bigquery/docs/reference/rest/v2/Job#Code_1